requirements.txt
pdf2docx
pdf2image
PyMuPDF
pytesseract
python-docx